dom/bindings/test/test_immutable_arraybufferview.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1952253
-->
<head>
  <meta charset="utf-8">
  <title>Test for immutable ArrayBufferViews</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1952253">Mozilla Bug 1952253</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">
    /* global TestFunctions */

    function checkThrowsImmutable(f) {
      let ex;
      try{
        f();
        ok(false, "Should have thrown!");
      } catch (e) {
        ex = e;
      }
      ok(ex.toString().includes("immutable ArrayBuffer or ArrayBufferView"), "Got exception: " + ex);
    }

    add_task(async function test_immutable_arraybuffer_views() {
      await SpecialPowers.pushPrefEnv({set: [["dom.expose_test_interfaces", true]]});
      let test = new TestFunctions();

      let ab = new ArrayBuffer(16).transferToImmutable();
      checkThrowsImmutable(() => test.testNotAllowShared(ab));
      checkThrowsImmutable(() => test.testAllowShared(ab));
      checkThrowsImmutable(() => test.testDictWithAllowShared({arrayBuffer: ab}));
      checkThrowsImmutable(() => test.testUnionOfBufferSource(ab));
      checkThrowsImmutable(() => { test.arrayBuffer = ab; });
      checkThrowsImmutable(() => { test.allowSharedArrayBuffer = ab; });
      checkThrowsImmutable(() => { test.sequenceOfArrayBuffer = [ab]; });
      checkThrowsImmutable(() => { test.sequenceOfAllowSharedArrayBuffer = [ab]; });

      let ta = new Int8Array(ab);
      checkThrowsImmutable(() => test.testNotAllowShared(ta));
      checkThrowsImmutable(() => test.testAllowShared(ta));
      checkThrowsImmutable(() => test.testDictWithAllowShared({arrayBufferView: ta}));
      checkThrowsImmutable(() => test.testUnionOfBufferSource(ta));
      checkThrowsImmutable(() => { test.arrayBufferView = ta; });
      checkThrowsImmutable(() => { test.allowSharedArrayBufferView = ta; });
      checkThrowsImmutable(() => { test.sequenceOfArrayBufferView = [ta]; });
      checkThrowsImmutable(() => { test.sequenceOfAllowSharedArrayBufferView = [ta]; });

      let dv = new DataView(ab);
      checkThrowsImmutable(() => test.testNotAllowShared(dv));
      checkThrowsImmutable(() => test.testAllowShared(dv));
      checkThrowsImmutable(() => test.testDictWithAllowShared({arrayBufferView: dv}));
      checkThrowsImmutable(() => test.testUnionOfBufferSource(dv));
      checkThrowsImmutable(() => { test.arrayBufferView = dv; });
      checkThrowsImmutable(() => { test.allowSharedArrayBufferView = dv; });
      checkThrowsImmutable(() => { test.sequenceOfArrayBufferView = [dv]; });
      checkThrowsImmutable(() => { test.sequenceOfAllowSharedArrayBufferView = [dv]; });
    });
  </script>
</body>
</html>
